Working on commands

+++ commands/uim/commands/classes/commands/caches/cleargroup.d
@@ -41,7 +41,7 @@ class DCacheClearGroupCommand : DCommand {
   }
 
   // Clears the cache group
-  override int execute(Json[string] arguments, IConsole aConsole = null) {
+  override ulong execute(Json[string] arguments, IConsole aConsole = null) {
     auto anGroup = to!string(commandArguments.getArgument("group"));
     try {
       anGroupConfigs = Cache.groupConfigs(anGroup);
