+++ commands/uim/commands/classes/commands/server.d
@@ -72,7 +72,7 @@ return false;
         }
     }
 
-  overwrite override int execute(Json[string] arguments, IConsole aConsole = null) {
+  overwrite override ulong execute(Json[string] arguments, IConsole aConsole = null) {
         this.startup(commandArguments,  aConsoleIo);
         DBinary = to!string(enviroment("D", "d"));
         string commandText = "%s -S %s:%d -t %s"
